Data-Pipeline/dags update/Chunk_Embeddings/requirements.txt
langchain
langchain-openai
langchain-pinecone
pinecone-client
openai